pyproject.toml
[project]
name = "mandarin-tutor-api"
version = "0.1.0"
description = ""
authors = [
    {name = "Jörg Besier",email = "joerg.besier@jbvv-digital.de"}
]
readme = "README.md"
requires-python = "^3.10"
dependencies = [
    "uvicorn[standard] (>=0.34.3,<0.35.0)",
    "fastapi (>=0.115.12,<0.116.0)",
    "jinja2 (>=3.1.6,<4.0.0)",
    "gspread (>=6.2.1,<7.0.0)",
    "google-auth (>=2.40.3,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "langchain-core (>=0.3.64,<0.4.0)",
    "langchain-openai (>=0.3.21,<0.4.0)",
    "openai (>=1.84.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.0"
pytest-asyncio = "^1.0.0"

[tool.pytest.ini_options]
markers = [
    "integration: tests that hit live OpenAI endpoints; require OPENAI_API_KEY",
]